move methods for mapping from repo in user_mapper
