+++ src/data_access/repositories/user_repository.py
@@ -6,47 +6,8 @@ from sqlalchemy.ext.asyncio import AsyncSession
 
 from src.apps.user.domain.entities import User
 from src.apps.user.repositories import IUserRepository
+from src.data_access.mappers.user_mapper import UserMapper
 from src.data_access.models import UserModel
-
-
-def domain_to_model(user: User) -> UserModel:
-    return UserModel(
-        first_name=user.first_name,
-        last_name=user.last_name,
-        email=user.email,
-        password=user.password,
-        username_tg=user.username_tg,
-        nick_tg=user.nick_tg,
-        nick_gmeet=user.nick_gmeet,
-        nick_gitlab=user.nick_gitlab,
-        nick_github=user.nick_github,
-        avatar=user.avatar,
-        is_superuser=user.is_superuser,
-        is_active=user.is_active,
-        created_at=user.created_at,
-        updated_at=user.updated_at,
-        id=user.id,
-    )
-
-
-def model_to_domain(user_model: UserModel) -> User:
-    return User(
-        first_name=user_model.first_name,
-        last_name=user_model.last_name,
-        email=user_model.email,
-        password=user_model.password,
-        username_tg=user_model.username_tg,
-        nick_tg=user_model.nick_tg,
-        nick_gmeet=user_model.nick_gmeet,
-        nick_gitlab=user_model.nick_gitlab,
-        nick_github=user_model.nick_github,
-        avatar=user_model.avatar,
-        is_superuser=user_model.is_superuser,
-        is_active=user_model.is_active,
-        _created_at=user_model.created_at,
-        _updated_at=user_model.updated_at,
-        _id=user_model.id,
-    )
 
 
 class UserRepository(IUserRepository):
@@ -55,7 +16,7 @@ class UserRepository(IUserRepository):
         self.model = UserModel
 
     async def save(self, user: User) -> None:
-        sql_user = domain_to_model(user)
+        sql_user = UserMapper.domain_to_model(user)
         user._created_at = datetime.datetime.now(datetime.timezone.utc)
         user._updated_at = datetime.datetime.now(datetime.timezone.utc)
         self._session.add(sql_user)
@@ -64,7 +25,7 @@ class UserRepository(IUserRepository):
         query = select(self.model).where(self.model.email == email.lower())
         result = await self._session.execute(query)
         sql_user = result.scalar_one_or_none()
-        return model_to_domain(sql_user) if sql_user else None
+        return UserMapper.model_to_domain(sql_user) if sql_user else None
 
     async def find_all_users(
         self, limit: int, offset: int, sort_by: Optional[str] = None, sort_order: str = 'asc'
@@ -80,7 +41,7 @@ class UserRepository(IUserRepository):
         query = query.offset(offset).limit(limit)
         result = await self._session.execute(query)
         sql_users = result.scalars().all()
-        return [model_to_domain(sql_user) for sql_user in sql_users]
+        return [UserMapper.model_to_domain(sql_user) for sql_user in sql_users]
 
     async def delete(self, email: str) -> None:
         query = select(self.model).where(self.model.email == email)
